[ui] #22 체크 컴포넌트 결정 후 수정
