+++ src/user/components/CheckBox/CheckBox.css
@@ -1,30 +1,49 @@
-/* 
-// NOTE 체크박스 체크 전
-*/
-.custom-checkbox-before {
-    width: 26px;
-    height: 26px;
-    padding: 0;
-    border: 2px solid darkgray;
-    border-radius: 5px;
-    background-color: #fff;
+.label {
+    display: inline-block;
+    position: relative;
+    padding-left: 54px;
+    cursor: pointer;
+    font-size: 15px;
 }
 
-/* 
-// NOTE 체크박스 체크 전 체크 svg 색상
-*/
-.custom-checkbox-before path {
-    fill: darkgray
+.label input {
+    position: absolute;
+    opacity: 0;
+    cursor: pointer;
 }
 
-/* 
-// NOTE 체크박스 체크 후
-*/
-.custom-checkbox-after {
-    width: 26px;
-    height: 26px;
-    padding: 0;
-    border: 2px solid #FFD465;
-    border-radius: 5px;
-    background-color: #FFD465;
+.checkmark {
+    position: absolute;
+    top: -1px;
+    left: 15px;
+    height: 26px; /* 체크박스 크기 조절 */
+    width: 26px; /* 체크박스 크기 조절 */
+    background-color: #ffffff;
+    border: 1px solid #9c9c9c;
+    border-radius: 4px;
+}
+
+.label input:checked + .checkmark {
+    background-color: #FFD465; /* 체크된 경우 배경색 변경 */
+    border: 0; /* 체크된 경우 테두리 색 변경 */
+}
+
+.checkmark:after {
+    content: "";
+    position: absolute;
+    display: none;
+}
+
+.label input:checked + .checkmark:after {
+    display: block;
+}
+
+.label .checkmark:after {
+    left: 9px;
+    top: 5px;
+    width: 5px;
+    height: 10px;
+    border: solid black;
+    border-width: 0 3px 3px 0;
+    transform: rotate(45deg);
 }
