+++ src/user/pages/Detail/Detail.css
@@ -530,91 +530,23 @@ button {
 }
 
 .secretArticle {
-    margin-left: 840px;
+    margin-left: 800px;
     font-weight: 500;
 }
 .public {
     display: inline-block;
     position: relative;
-    padding-left: 54px;
+    padding-left: 20px;
     cursor: pointer;
     font-size: 15px;
 }
-
-.public input {
-    position: absolute;
-    opacity: 0;
-    cursor: pointer;
-  }
-.checkmark {
-    position: absolute;
-    top: -1px;
-    left: 15px;
-    height: 26px; /* 체크박스 크기 조절 */
-    width: 26px; /* 체크박스 크기 조절 */
-    background-color: #ffffff;
-    border: 1px solid #9c9c9c;
-    border-radius: 4px;
-  }
-
-.public input:checked + .checkmark {
-    background-color: #FFD465; /* 체크된 경우 배경색 변경 */
-    border: 0; /* 체크된 경우 테두리 색 변경 */
-  }
-.checkmark:after {
-    content: "";
-    position: absolute;
-    display: none;
-  }
-
-.public input:checked + .checkmark:after {
-    display: block;
-  }
-
-.public .checkmark:after {
-    left: 9px;
-    top: 5px;
-    width: 5px;
-    height: 10px;
-    border: solid black;
-    border-width: 0 3px 3px 0;
-    transform: rotate(45deg);
-  }
 
 .private {
     display: inline-block;
     position: relative;
-    padding-left: 54px;
     cursor: pointer;
     font-size: 15px;
 }
-
-.private input {
-    position: absolute;
-    opacity: 0;
-    cursor: pointer;
-  }
-
-
-.private input:checked + .checkmark {
-    background-color: #FFD465; /* 체크된 경우 배경색 변경 */
-    border: 0; /* 체크된 경우 테두리 색 변경 */
-  }
-
-
-.private input:checked + .checkmark:after {
-    display: block;
-  }
-
-.private .checkmark:after {
-    left: 9px;
-    top: 5px;
-    width: 5px;
-    height: 10px;
-    border: solid black;
-    border-width: 0 3px 3px 0;
-    transform: rotate(45deg);
-  }
 
 .enrollButton {
     margin-left: 963px;
