+++ src/user/pages/Login/Login.css
@@ -88,13 +88,12 @@
   display: flex; 
   justify-content: space-between;
   align-items: center;
-  margin-top: 20px;
+  margin-top: 10px;
   margin-bottom: 38px;
   align-self: flex-start;
 }
   
-  
-.login-checkbox {
+/* .login-checkbox {
   width: 26px;
   height: 26px;
   border-radius: 5px;
@@ -116,6 +115,14 @@
   text-align: center;
   position: relative;
   font-size: 20px;
+} */
+
+.login-checkbox-label {
+  display: inline-block;
+  position: relative;
+  cursor: pointer;
+  font-size: 15px;
+  padding-top: 20px;
 }
   
 .login-btn {
