+++ src/user/pages/MyPage/MyActivity/WishList/WishList.css
@@ -2,6 +2,7 @@
     width: 1100px;
     margin: 0 auto;
     padding: 0 70px;
+    padding-bottom: 200px;
 }
 
 .wishList-title {
@@ -18,6 +19,12 @@
     display: flex;
     align-items: center;
     gap: 8px;
+}
+
+.wishList-checkbox {
+    display: flex;
+    align-items: center;
+    padding-bottom: 20px;
 }
 
 .wishList-count {
